Modified class names and created new class
Modified class names back to their original ones due to
CLI not being able to find entry points.

Also working on reformating the architecture of the classes to be
more maintainable.  Adding constructor for the retreive messages
which will take the arguments for sending up a message retrieval
as a dictionary.  Also working on making it so device parses the
arguments prior to sending them to retrieveMessage so all of that
original processing can be done at that level instead of duplicated
across the entire project for each action taken by the device.