Added validation checks in Device Receive
Parses the strings and files provided to the command line to verify
that they are what we are looking for prior to handing the items over to
RetrieveMessages.  Working on stripping down the original
retrieveMessages due to it becoming a monolithic class.  Moving
everything to RetrieveMessagesTest, eventually, and will only need to
worry about actuall retrieveing the messages. Device is going to handle
connection string and RSA cert validation.